fix bugs with product name textbox add check on correct product name

+++ Lab7/Form1.cs
@@ -1,3 +1,5 @@
+using System.Text.RegularExpressions;
+
 namespace Lab7
 {
     public partial class Form1 : Form
@@ -12,17 +14,36 @@ namespace Lab7
             onlineStore = new OnlineStore("Интернет магазин");
             onlineStore.ProductsWork += DisplayMessage;
         }
-
+        string namePattern = @"[а-яА-я]+[0-9]*";
+        bool CheckName(string name)
+        {
+            return Regex.IsMatch(name, namePattern);
+        }
         void DisplayMessage(string message) => MessageBox.Show(message, "Сообщение");
 
         private void button1_Click(object sender, EventArgs e)
         {
-            onlineStore.AddProduct(new Product(productName.Text, productPrice.Value));
+            if (CheckName(productName.Text))
+            {
+                onlineStore.AddProduct(new Product(productName.Text, productPrice.Value));
+            }
+            else
+            {
+                MessageBox.Show("Ошибка в назвнаии продукта", "Сообщение");
+            }
         }
 
         private void button1_Click_1(object sender, EventArgs e)
         {
-            onlineStore.DeleteProduct(new Product(productName.Text, productPrice.Value));
+            if (CheckName(productName.Text))
+            {
+                onlineStore.DeleteProduct(new Product(productName.Text, productPrice.Value));
+            }
+            else
+            {
+                MessageBox.Show("Ошибка в назвнаии продукта", "Сообщение");
+            }
+            
         }
     }
 }
